requirements/prod.txt
Django>=1.6.2,<=1.6.8
djangorestframework==2.4.3
django-rest-swagger>=0.2.0
gunicorn>=18.0
pysnmp>=4.2.5
pyyaml==3.11
